Rename thread package to threads
