+++ src/main/java/org/asamk/signal/Manager.java
@@ -67,8 +67,8 @@ import org.asamk.signal.storage.groups.GroupInfo;
 import org.asamk.signal.storage.groups.JsonGroupStore;
 import org.asamk.signal.storage.protocol.JsonIdentityKeyStore;
 import org.asamk.signal.storage.protocol.JsonSignalProtocolStore;
-import org.asamk.signal.storage.thread.JsonThreadStore;
-import org.asamk.signal.storage.thread.ThreadInfo;
+import org.asamk.signal.storage.threads.JsonThreadStore;
+import org.asamk.signal.storage.threads.ThreadInfo;
 import org.asamk.signal.util.Base64;
 import org.asamk.signal.util.Util;
 import org.whispersystems.libsignal.DuplicateMessageException;
