+++ src/main/java/org/asamk/signal/storage/threads/JsonThreadStore.java
@@ -1,4 +1,4 @@
-package org.asamk.signal.storage.thread;
+package org.asamk.signal.storage.threads;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonGenerator;
